add classes
